[2024-04-22] [정세영] 산출물 로그인 체크: 관리자거나 pm이 아닌경우 404에러 던짐

+++ src/main/java/com/ktdsuniversity/edu/pms/output/web/OutputController.java
@@ -24,15 +24,14 @@ import com.ktdsuniversity.edu.pms.output.service.OutputService;
 import com.ktdsuniversity.edu.pms.output.vo.OutputListVO;
 import com.ktdsuniversity.edu.pms.output.vo.OutputSearchVO;
 import com.ktdsuniversity.edu.pms.output.vo.OutputVO;
+import com.ktdsuniversity.edu.pms.project.dao.ProjectDao;
 import com.ktdsuniversity.edu.pms.project.service.ProjectService;
 import com.ktdsuniversity.edu.pms.project.vo.ProjectListVO;
 import com.ktdsuniversity.edu.pms.project.vo.ProjectTeammateVO;
-import com.ktdsuniversity.edu.pms.project.vo.ProjectVO;
 import com.ktdsuniversity.edu.pms.utils.AjaxResponse;
 import com.ktdsuniversity.edu.pms.utils.Validator;
 import com.ktdsuniversity.edu.pms.utils.Validator.Type;
 
-import ch.qos.logback.core.filter.Filter;
 
 @Controller
 public class OutputController {
@@ -42,6 +41,8 @@ public class OutputController {
 	private ProjectService projectService;
 	@Autowired
 	private CommonCodeService commonCodeService;
+	@Autowired
+	private ProjectDao projectDao;
 
 	@GetMapping("/output")
 	public String viewOutputList() {
@@ -52,13 +53,7 @@ public class OutputController {
 	public String viewOutputSearhList(@SessionAttribute("_LOGIN_USER_") EmployeeVO employeeVO,
 			@RequestParam String prjId, Model model, OutputSearchVO outputSearchVO) {
 		
-		if(! employeeVO.getAdmnCode().equals("301") )  {//관리자가 아닌경우
-			if(employeeVO.getAdmnCode().equals("PM")) {//pm인경우
-				outputSearchVO.setEmpId(employeeVO.getEmpId());
-			}else {
-				throw new PageNotFoundException();
-			}
-		}
+		this.checkAccess(employeeVO, prjId);
 		
 		ProjectListVO projectList = this.projectService.getAllProject();
 		projectList.setProjectList(
@@ -76,8 +71,9 @@ public class OutputController {
 	}
 
 	@GetMapping("/output/write")
-	public String viewCreateOutput(Model model) {
-//		TODO 파일 넣기
+	public String viewCreateOutput(@SessionAttribute("_LOGIN_USER_") EmployeeVO employeeVO,Model model) {
+		this.checkAccess(employeeVO);
+		
 		ProjectListVO projectList = this.projectService.getAllProject();
 		projectList.setProjectList(
 				projectList.getProjectList().stream().filter((project) -> project.getOutYn().equals("Y")).toList());
@@ -89,7 +85,9 @@ public class OutputController {
 	}
 
 	@PostMapping("/output/write")
-	public String createOutput(@RequestParam MultipartFile file, OutputVO outputVO, Model model) {
+	public String createOutput(@SessionAttribute("_LOGIN_USER_") EmployeeVO employeeVO,@RequestParam MultipartFile file, OutputVO outputVO, Model model) {
+		this.checkAccess(employeeVO, outputVO.getPrjId());
+		
 		Validator<OutputVO> validator = new Validator<>(outputVO);
 		validator.add("outTtl", Type.NOT_EMPTY, "제목은 필수 입력값입니다").add("outType", Type.NOT_EMPTY, "산출물 타입은 필수 입력값입니다")
 				.add("prjId", Type.NOT_EMPTY, "올바르지 않은 프로젝트에서 생성했습니다.").start();
@@ -101,8 +99,9 @@ public class OutputController {
 	}
 
 	@GetMapping("output/downloadFile/{outId}")
-	public ResponseEntity<Resource> fileDownload(@PathVariable String outId) {
-
+	public ResponseEntity<Resource> fileDownload(@SessionAttribute("_LOGIN_USER_") EmployeeVO employeeVO,@PathVariable String outId) {
+		this.checkAccess(employeeVO);
+		
 		OutputVO Output = this.outputService.getOneOutput(outId);
 
 		return this.outputService.getDownloadFile(Output);
@@ -110,7 +109,9 @@ public class OutputController {
 	}
 
 	@GetMapping("/output/modify/{outId}")
-	public String viewModifyOutputPage(@PathVariable String outId, Model model) {
+	public String viewModifyOutputPage(@SessionAttribute("_LOGIN_USER_") EmployeeVO employeeVO,@PathVariable String outId, Model model) {
+		this.checkAccess(employeeVO);
+		
 		ProjectListVO projectList = this.projectService.getAllProject();
 		List<CommonCodeVO> outputType = this.commonCodeService.getAllCommonCodeListByPId("1000");
 		OutputVO output = this.outputService.getOneOutput(outId);
@@ -123,8 +124,9 @@ public class OutputController {
 	}
 
 	@PostMapping("/output/modify/{outId}")
-	public String ModifyOutputPage(@PathVariable String outId, @RequestParam MultipartFile file, OutputVO outputVO) {
-
+	public String ModifyOutputPage(@SessionAttribute("_LOGIN_USER_") EmployeeVO employeeVO,@PathVariable String outId, @RequestParam MultipartFile file, OutputVO outputVO) {
+		this.checkAccess(employeeVO, outputVO.getPrjId());
+		
 		Validator<OutputVO> validator = new Validator<>(outputVO);
 		validator.add("outTtl", Type.NOT_EMPTY, "제목은 필수 입력값입니다").add("outType", Type.NOT_EMPTY, "산출물 타입은 필수 입력값입니다")
 				.add("prjId", Type.NOT_EMPTY, "올바르지 않은 프로젝트에서 생성했습니다.").start();
@@ -134,12 +136,47 @@ public class OutputController {
 	}
 
 	@GetMapping("/output/delete/{outId}")
-	public String deleteOutputment(@PathVariable String outId, @RequestParam String prjId) {
+	public String deleteOutputment(@SessionAttribute("_LOGIN_USER_") EmployeeVO employeeVO,@PathVariable String outId, @RequestParam String prjId) {
+		this.checkAccess(employeeVO, prjId);
 
 		boolean isSuccess = this.outputService.deleteOneOutput(outId);
 
 		return "redirect:/output/search?prjId=" + prjId;
 
 	}
-
+	
+	private void checkAccess(EmployeeVO employeeVO, String prjId) {
+		ProjectTeammateVO pmVO =this.projectDao.findPmByProjectId(prjId);
+		if(! employeeVO.getAdmnCode().equals("301") )  {//관리자가 아닌경우
+			if(pmVO != null) {//프로젝트 아이디가 주어진경우
+				if(pmVO.getTmId().equals(employeeVO.getEmpId())) {//pm인경우
+				}else {//pm이 아닌경우
+					throw new PageNotFoundException();
+				}
+			}else{//프로젝트 아이디가 안주어진 경우
+				List<ProjectTeammateVO>  tmList = this.projectService.getAllProjectTeammate()
+				.stream()
+				.filter(tm -> tm.getTmId().equals(employeeVO.getEmpId()))
+				.filter(tm -> tm.getRole().equals("PM")).toList();
+				
+				if(tmList ==null || tmList.isEmpty()) {//PM을 맏은 포지션이 없다면
+					throw new PageNotFoundException();
+				}else {}//pm을 맞은 포지션이 있다면
+			}
+		}
+	}
+		
+		private void checkAccess(EmployeeVO employeeVO) {
+			if(! employeeVO.getAdmnCode().equals("301") )  {//관리자가 아닌경우
+					//프로젝트 아이디가 안주어진 경우
+				List<ProjectTeammateVO>  tmList = this.projectService.getAllProjectTeammate()
+				.stream()
+				.filter(tm -> tm.getTmId().equals(employeeVO.getEmpId()))
+				.filter(tm -> tm.getRole().equals("PM")).toList();
+	
+				if(tmList ==null || tmList.isEmpty()) {//PM을 맏은 포지션이 없다면
+					throw new PageNotFoundException();
+				}else {}//pm을 맞은 포지션이 있다면
+		}
+	}
 }
